+++ src/renderer/App.tsx
@@ -4,7 +4,7 @@ import Voice from './Voice';
 import Menu from './Menu';
 import { ipcRenderer, remote } from 'electron';
 import { AmongUsState } from '../common/AmongUsState';
-import Settings, { settingsReducer } from './Settings';
+import Settings, { settingsReducer } from './settings/Settings';
 import { GameStateContext, SettingsContext } from './contexts';
 
 let appVersion = '';
