+++ src/renderer/settings/MicrophoneSoundBar.tsx
